requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
python-dotenv>=1.0.0
openai>=1.0.0
psycopg2-binary==2.9.9
dj-database-url>=2.0.0
djangorestframework-simplejwt>=5.2.2 